Add tests and add better documentation and example
