+++ python/ray/tune/tests/test_tune_restore.py
@@ -10,7 +10,8 @@ import unittest
 
 import ray
 from ray import tune
-from ray.tune.util import recursive_fnmatch, validate_save_restore
+from ray.tests.util import recursive_fnmatch
+from ray.tune.util import validate_save_restore
 from ray.rllib import _register_all
