+++ python/ray/tune/util.py
@@ -213,18 +213,6 @@ def _from_pinnable(obj):
     return obj[0]
 
 
-def recursive_fnmatch(dirpath, pattern):
-    """Looks at a file directory subtree for a filename pattern.
-
-    Similar to glob.glob(..., recursive=True) but also supports 2.7
-    """
-    matches = []
-    for root, dirnames, filenames in os.walk(dirpath):
-        for filename in fnmatch.filter(filenames, pattern):
-            matches.append(os.path.join(root, filename))
-    return matches
-
-
 def validate_save_restore(trainable_cls, config=None, use_object_store=False):
     """Helper method to check if your Trainable class will resume correctly.
